+++ src/backend/sqlite/foreign_key.rs
@@ -1,17 +1,17 @@
 use super::*;
 
 impl ForeignKeyBuilder for SqliteQueryBuilder {
-    fn prepare_foreign_key_drop_statement(
+    fn prepare_foreign_key_drop_statement_internal(
         &self,
         drop: &ForeignKeyDropStatement,
         sql: &mut SqlWriter,
+        mode: Mode,
     ) {
-        write!(sql, "ALTER TABLE ").unwrap();
-        if let Some(table) = &drop.table {
-            table.prepare(sql, self.quote());
+        if mode != Mode::Creation {
+            panic!("Sqlite does not support modification of foreign key constraints to existing tables");
         }
 
-        write!(sql, " DROP FOREIGN KEY ").unwrap();
+        write!(sql, "DROP FOREIGN KEY ").unwrap();
         if let Some(name) = &drop.foreign_key.name {
             write!(sql, "`{}`", name).unwrap();
         }
@@ -21,9 +21,9 @@ impl ForeignKeyBuilder for SqliteQueryBuilder {
         &self,
         create: &ForeignKeyCreateStatement,
         sql: &mut SqlWriter,
-        inside_table_creation: bool,
+        mode: Mode,
     ) {
-        if !inside_table_creation {
+        if mode != Mode::Creation {
             panic!("Sqlite does not support modification of foreign key constraints to existing tables");
         }