+++ src/lib.rs
@@ -47,7 +47,7 @@
 //!
 //! ### Feature flags
 //!
-//! Macro: `derive`
+//! Macro: `derive` `attr`
 //!
 //! Async support: `thread-safe` (use `Arc` inplace of `Rc`)
 //!
@@ -183,12 +183,13 @@
 //! ```
 //!
 //! If you're okay with running another procedural macro, you can activate
-//! the `derive` feature on the crate to save you some boilerplate.
+//! the `derive` or `attr` feature on the crate to save you some boilerplate.
 //! For more usage information, look at
-//! [the derive examples](https://github.com/SeaQL/sea-query/tree/master/sea-query-derive/tests/pass).
+//! [the derive examples](https://github.com/SeaQL/sea-query/tree/master/sea-query-derive/tests/pass)
+//! or [the attribute examples](https://github.com/SeaQL/sea-query/tree/master/sea-query-attr/tests/pass).
 //!
 //! ```rust
-//! # #[cfg(feature = "derive")]
+//! #[cfg(feature = "derive")]
 //! use sea_query::Iden;
 //!
 //! // This will implement Iden exactly as shown above
@@ -203,6 +204,32 @@
 //! struct Glyph;
 //! assert_eq!(Glyph.to_string(), "glyph");
 //! ```
+//!
+//! ```rust
+//! #[cfg(feature = "attr")]
+//! # fn test() {
+//! use sea_query::{Iden, enum_def};
+//!
+//! #[enum_def]
+//! struct Character {
+//!   pub foo: u64,
+//! }
+//!
+//! // It generates the following along with Iden impl
+//! # let not_real = || {
+//! enum CharacterIden {
+//!     Table,
+//!     Foo,
+//! }
+//! # };
+//!
+//! assert_eq!(CharacterIden::Table.to_string(), "character");
+//! assert_eq!(CharacterIden::Foo.to_string(), "foo");
+//! # }
+//! # #[cfg(feature = "attr")]
+//! # test();
+//! ```
+//!
 //!
 //! ### Expression
 //!
@@ -751,5 +778,8 @@ pub use value::*;
 
 #[cfg(feature = "derive")]
 pub use sea_query_derive::Iden;
+
+#[cfg(feature = "attr")]
+pub use sea_query_attr::enum_def;
 
 pub use sea_query_driver::*;